feat: implement user login state management; refactor home screen and update routing logic
