+++ lib/core/helpers/constants.dart
@@ -2,4 +2,5 @@ bool isLoggedInUser = false;
 
 class SharedPrefKeys {
   static const String userToken = 'userToken';
+  bool isLoggedInUser = false;
 }
