+++ lib/doctor_app.dart
@@ -1,3 +1,4 @@
+import 'core/helpers/constants.dart';
 import 'core/routing/app_route.dart';
 import 'core/routing/routes.dart';
 import 'package:flutter/material.dart';
@@ -9,14 +10,14 @@ class DoctorApp extends StatelessWidget {
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
-    return const ScreenUtilInit(
-      designSize: Size(375, 812),
+    return ScreenUtilInit(
+      designSize: const Size(375, 812),
       minTextAdapt: true,
       splitScreenMode: true,
       child: MaterialApp(
         debugShowCheckedModeBanner: false,
 
-        initialRoute: Routes.loginScreen,
+        initialRoute: isLoggedInUser ? Routes.homeScreen : Routes.loginScreen,
         onGenerateRoute: AppRoute.generateRoute,
       ),
     );
